ai_engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
torch>=2.3.0
Pillow>=10.0.0
requests>=2.32.0

# YOLOv10 implementation (expects yolov10s.pt weights to be available)
ultralytics

# Ethereum signing & env management
eth-account>=0.10.0
python-dotenv>=1.0.1
eth-abi>=4.0.0b2


